moving dag_util to roger module
